requirements.txt
python-dotenv
ollama
requests
speechrecognition
pyttsx3
pyaudio
keyboard
pocketsphinx
langchain 
langchain-community
langchain-ollama
pyautogui
pytesseract
pygetwindow
Pillow
winshell
pywin32
psutil